requirements.txt
google-api-python-client==2.94.0
pandas==2.1.2
boto3==1.28.58
python-dotenv==1.0.0
